pyproject.toml
[project]
name = "centuria"
version = "0.1.0"
description = "Generate LLM personas to model groups of humans and survey them"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.40.0",
    "pydantic>=2.0",
    "pandas>=2.0",
    "jupyter>=1.0",
    "jupyterlab>=4.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pypdf>=4.0",
    "tiktoken>=0.7.0",
    "plotly>=5.0",
    "python-dotenv>=1.0",
    "aiohttp>=3.9",
    "diskcache>=5.6",
    "tenacity>=8.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/centuria"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
